pyproject.toml
[project]
name = "simple-taskpy-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=2.0.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
dataset = {cmd="python scripts/download_dataset.py", help="download the dataset and place it in to data/ folder."}
